The lint-and-test workflow fails only if the automated tests for the versions used in the image fail.
The workflow should pass if the tests pass when running under the same versons of Elixir and Erlang/OTP as the image.
The matrix of tests is useful to see if we can use newer versions, but these tests should not make the workflow fail.

As some of the source files are not formatted, the pseudo-linting fails.
When we format the files, we should change the psudo-linting to fail the workflow, so we can keep the files formatted.